+++ src/commands/firestore-backups-delete.ts
@@ -12,6 +12,7 @@ import { FirebaseError } from "../error";
 export const command = new Command("firestore:backups:delete <backup>")
   .description("Delete a backup under your Cloud Firestore database.")
   .option("--force", "Attempt to delete backup without prompting for confirmation.")
+  .firebaseNotRequired()
   .before(requirePermissions, ["datastore.backups.delete"])
   .before(warnEmulatorNotSupported, Emulators.FIRESTORE)
   .action(async (backupName: string, options: FirestoreOptions) => {