+++ src/commands/serve.ts
@@ -24,6 +24,7 @@ function filterOnly(list: string[], only = ""): string[] {
 
 export const command = new Command("serve")
   .description("start a local server for your static assets")
+  .firebaseNotRequired()
   .option("-p, --port <port>", "the port on which to listen (default: 5000)", 5000)
   .option("-o, --host <host>", "the host on which to listen (default: localhost)", "localhost")
   .option(