+++ CHANGELOG.md
@@ -1,1 +1,2 @@
+- Improved error messaging when using a project that does not have Firebase enabled.
 - Changes default CF3 runtime to nodejs22 (#8037)